Add support for breakpoints to stagger animation
