+++ src/progress_bar/utils.js
@@ -1,0 +1,3 @@
+export function clamp(n, min, max) {
+  return Math.max(min, Math.min(n, max));
+}
